requirements.txt

Flask>=3.0.0
python-dotenv>=1.0.1
pandas>=2.2.2
numpy>=1.26.4
yfinance>=0.2.40
plotly>=5.22.0
twilio>=9.0.0
APScheduler>=3.10.4
email-validator>=2.2.0
requests>=2.32.3
tenacity>=8.4.2
